backend/requirements-dev.txt
-r requirements.txt
pytest==8.*
pytest-django==4.*
ruff==0.9.*
